LayoutTests/accessibility/mac/AOM-bool-properties-expected.txt

outside of dialog

dialog
This tests boolean type Accessibility Object Model properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Test AOM boolean properties can be cleared.
PASS axNode.boolAttributeValue('AXARIAAtomic') is true
PASS node.accessibleNode.atomic is null
PASS axNode.boolAttributeValue('AXARIAAtomic') is false
PASS node.accessibleNode.atomic is null
PASS axNode.boolAttributeValue('AXARIAAtomic') is true

Busy support
PASS axNode.boolAttributeValue('AXElementBusy') is false
PASS axNode.boolAttributeValue('AXElementBusy') is true

Disabled support
PASS axNode.isEnabled is true
PASS axNode.isEnabled is false

Expanded support
PASS axNode.isExpanded is false
PASS axNode.isExpanded is true

Hidden support
PASS axNode.isIgnored is false
PASS axNode == null || axNode == undefined is true

Modal support
PASS dialogBackgroundAccessible() is true
PASS dialogBackgroundAccessible() is false

Multiline support
PASS axNode.boolAttributeValue('AXIsMultiline') is false
PASS axNode.boolAttributeValue('AXIsMultiline') is true

Multiselectable support
PASS axNode.isMultiSelectable is false
PASS axNode.isMultiSelectable is true

ReadOnly support
PASS axNode.stringAttributeValue('AXReadOnlyValue') is "false"
PASS axNode.stringAttributeValue('AXReadOnlyValue') is "true"

Required support
PASS axNode.isRequired is false
PASS axNode.isRequired is true

Selected support
PASS axNode.isSelected is false
PASS axNode.isSelected is true
